sensors/requirements.txt
pyserial == 3.4
requests